refactor(work_log): split work_log_page into per-widget files
work_log/widgets/:
- date_header.dart         DateHeader (today/YYYY-MM-DD selector text)
- work_log_card.dart       WorkLogCard (the swipe-to-delete card row)
- add_exercise_fab.dart    AddExerciseFab (bottom-right + add button)
- add_exercise_dialog.dart AddExerciseDialog (the picker dialog used
                            by the FAB)

WorkLogCard reuses the shared DeleteActionPane added in the previous
commit. Page file shrinks from ~420 lines to ~145 lines.
