+++ test/presentation/pages/work_log/work_log_page_test.dart
@@ -7,13 +7,13 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:workout_log/domain/models/body_part.dart';
 import 'package:workout_log/domain/models/exercise.dart';
 import 'package:workout_log/domain/models/work_log.dart';
-import 'package:workout_log/presentation/pages/exercise_detail_page.dart';
-import 'package:workout_log/presentation/pages/work_log_page.dart';
+import 'package:workout_log/presentation/pages/exercise_detail/exercise_detail_page.dart';
+import 'package:workout_log/presentation/pages/work_log/work_log_page.dart';
 import 'package:workout_log/presentation/providers/data_providers.dart';
 import 'package:workout_log/presentation/providers/selected_date_provider.dart';
 
-import '../helpers/test_app.dart';
-import '../test_helper.dart';
+import '../../../helpers/test_app.dart';
+import '../../../test_helper.dart';
 
 WorkLog _seedWorkLog(String name, DateTime date,
     {Set<BodyPart> bodyParts = const {BodyPart.chest}}) {
